requirements.txt
torch==2.1.0
torchvision==0.16.0
pytorch-lightning==2.1.0
einops==0.7.0
wandb==0.16.0
hydra-core==1.3.2
albumentations==1.3.1
matplotlib==3.7.2
seaborn==0.13.0
torchcam==0.3.5
scikit-learn==1.3.2
tqdm==4.66.1
ipywidgets==8.1.1
